Move resolveProperties function
