+++ template_engine/plugins.ts
@@ -9,6 +9,10 @@ import {
 import {
   instanceOfRef,
   StateStack,
+  Template,
+  EvaluationTemplate,
+  HasAttrTemplate,
+  GroupTemplate,
   CustomTemplate,
   CustomElementTemplate,
   Cache,
@@ -148,5 +152,50 @@ export let snippetPlugin = {
     }
     let tree = evaluate(snippet.template, snippet.restack([...stack, { attrs }, attrs]), cache) as VirtualTree
     return tree.children
+  }
+}
+
+export function resolveProperties(
+  template: CustomElementTemplate,
+  stack: StateStack,
+  cache: Cache,
+  ve: VirtualElement | RealTarget
+)
+{
+  let values = [] as Array<Template | string>
+  let contents = [] as Array<[string, Template]>
+  let children = (template.children || [])
+    .map(child => {
+      if (!(typeof child === 'string')) {
+        let temp = child as HasAttrTemplate
+        if (temp.attrs) {
+          if (temp.attrs['@as']) {
+            contents.push([temp.attrs['@as'] as string, temp])
+            return []
+          } else if(temp.attrs.slot) {
+            return [evaluate(child, stack, cache) as string | VirtualElement | number]
+          }
+        }
+      }
+      values.push(child)
+      return []
+    })
+    .reduce((ary, values) => { // flatMap
+      ary.push(...values)
+      return ary
+    }, [])
+  if (values.length) {
+    contents.push(['content', { type: 'group', children: values } as GroupTemplate])
+  }
+  if (contents.length) {
+    if (!ve.attrs) {
+      ve.attrs = {}
+    }
+    contents.forEach(([name, value]) => {
+      (ve.attrs as Record<string, unknown | Template>)[name] = { type: 'evaluation', value, stack } as EvaluationTemplate
+    })
+  }
+  if (children.length) {
+    ve.children = children
   }
 }
