+++ web_components/plugins.ts
@@ -10,17 +10,14 @@ import {
 import { RealTarget, VirtualElement } from '../virtual_dom/types.ts'
 import {
   StateStack,
-  Template,
-  EvaluationTemplate,
-  HasAttrTemplate,
-  GroupTemplate,
   CustomElementTemplate,
   CustomTemplate,
   Cache,
   EvaluatePlugin
 } from '../template_engine/types.ts'
 import { ComponentElement, componentElementTag } from './element.ts'
-import { evaluate, evaluateAttrs, evaluateChildren } from '../template_engine/evaluate.ts'
+import { evaluateAttrs, evaluateChildren } from '../template_engine/evaluate.ts'
+import { resolveProperties } from '../template_engine/plugins.ts'
 import { isPrimitive } from '../template_engine/is_primitive.ts'
 import { pickup } from '../template_engine/pickup.ts'
 
@@ -219,49 +216,3 @@ export let specialTagPlugin = {
   }
 } as EvaluatePlugin
 
-
-    // Resolve properties
-function resolveProperties(
-  template: CustomElementTemplate | ComponentTemplate,
-  stack: StateStack,
-  cache: Cache,
-  ve: VirtualElement | RealTarget
-)
-{
-  let values = [] as Array<Template | string>
-  let contents = [] as Array<[string, Template]>
-  let children = (template.children || [])
-    .map(child => {
-      if (!(typeof child === 'string')) {
-        let temp = child as HasAttrTemplate
-        if (temp.attrs) {
-          if (temp.attrs['@as']) {
-            contents.push([temp.attrs['@as'] as string, temp])
-            return []
-          } else if(temp.attrs.slot) {
-            return [evaluate(child, stack, cache) as string | VirtualElement | number]
-          }
-        }
-      }
-      values.push(child)
-      return []
-    })
-    .reduce((ary, values) => { // flatMap
-      ary.push(...values)
-      return ary
-    }, [])
-  if (values.length) {
-    contents.push(['content', { type: 'group', children: values } as GroupTemplate])
-  }
-  if (contents.length) {
-    if (!ve.attrs) {
-      ve.attrs = {}
-    }
-    contents.forEach(([name, value]) => {
-      (ve.attrs as Record<string, unknown | Template>)[name] = { type: 'evaluation', value, stack } as EvaluationTemplate
-    })
-  }
-  if (children.length) {
-    ve.children = children
-  }
-}
